Found an interesting bug that accidently limited the number of elements that could be used in a single dimension of an array. Not sure exactly where the cut off was but I'm guessing that it was about 2^23 elements. Some ill conceived casts to floats were the culprit. I've fixed all occurances of this. Note arrays could be much larger than this but no single dimension could have more than 2^23 unless it was and odd size dimension in which case larger dimensions were supported (don't ask :) ).
	-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6285 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/NclFile.c
@@ -1125,7 +1125,7 @@ int vtype;
 					stride[sel->dim_num] = 1;
 					tmpf = 1;
 				}
-				n_elem =(int)(fabs(((float)(finish[sel->dim_num] - start[sel->dim_num]))) /tmpf) + 1;
+				n_elem =(int)(fabs(((double)(finish[sel->dim_num] - start[sel->dim_num]))) /tmpf) + 1;
 				if((sel->u.sub.start > thefile->file.file_dim_info[thefile->file.var_info[index]->file_dim_num[sel->dim_num]]->dim_size-1)||(sel->u.sub.start < 0)) {
 					NhlPError(NhlFATAL,NhlEUNKNOWN,"Subscript out of range, error in subscript #%d",i);
 					return(NULL);
@@ -2839,7 +2839,7 @@ int type;
 							stride[sel->dim_num] = 1;
 							tmpf = 1;
 						}
-						n_elem = (int)(fabs(((float)(finish[sel->dim_num] -start[sel->dim_num])))/tmpf) + 1;
+						n_elem = (int)(fabs(((double)(finish[sel->dim_num] -start[sel->dim_num])))/tmpf) + 1;
 
 						if((sel->u.sub.start > thefile->file.file_dim_info[thefile->file.var_info[index]->file_dim_num[sel->dim_num]]->dim_size-1 )||(sel->u.sub.start < 0)) {
 							if(!( thefile->file.file_dim_info[ thefile->file.var_info[index]->file_dim_num[sel->dim_num]]->is_unlimited)||(sel->u.sub.start < 0)) {
@@ -3693,9 +3693,9 @@ struct _NclSelectionRecord *rhs_sel_ptr;
 						break;
 					default:
 						if(lhs_sel_ptr->selection[i].u.sub.finish < lhs_sel_ptr->selection[i].u.sub.start) {
-							lhs_n_elem = (int)(((float)(lhs_sel_ptr->selection[i].u.sub.start - lhs_sel_ptr->selection[i].u.sub.finish))/(float)fabs(((float)lhs_sel_ptr->selection[i].u.sub.stride))) + 1;
+							lhs_n_elem = (int)(((double)(lhs_sel_ptr->selection[i].u.sub.start - lhs_sel_ptr->selection[i].u.sub.finish))/(double)fabs(((double)lhs_sel_ptr->selection[i].u.sub.stride))) + 1;
 						} else {
-							lhs_n_elem = (int)(((float)(lhs_sel_ptr->selection[i].u.sub.finish - lhs_sel_ptr->selection[i].u.sub.start))/(float)fabs(((float)lhs_sel_ptr->selection[i].u.sub.stride))) + 1;
+							lhs_n_elem = (int)(((double)(lhs_sel_ptr->selection[i].u.sub.finish - lhs_sel_ptr->selection[i].u.sub.start))/(double)fabs(((double)lhs_sel_ptr->selection[i].u.sub.stride))) + 1;
 						}
 						break;
 					}
