+++ ni/src/ncl/NclMultiDVallistData.c
@@ -1,6 +1,6 @@
 
 /*
- *      $Id: NclMultiDVallistData.c,v 1.2 2000-03-10 20:33:57 ethan Exp $
+ *      $Id: NclMultiDVallistData.c,v 1.3 2001-07-16 19:48:22 ethan Exp $
  */
 /************************************************************************
 *									*
@@ -147,10 +147,10 @@ static struct _NclDataRec *MultiDVal_list_ReadSection
                                         sel_ptr->u.sub.stride = 1;
                                 }
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.start 
 					- sel_ptr->u.sub.finish))
-					/(float)fabs(((float)sel_ptr->u.sub.stride))) + 1;
+					/(double)fabs(((double)sel_ptr->u.sub.stride))) + 1;
 
 /*
 * Need to be able to determine which type of comparision < or > is needed to
@@ -177,10 +177,10 @@ static struct _NclDataRec *MultiDVal_list_ReadSection
                                 }
 
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.finish 
 					- sel_ptr->u.sub.start))
-					/((float)sel_ptr->u.sub.stride)) + 1;
+					/((double)sel_ptr->u.sub.stride)) + 1;
 
 				
 				if(sel_ptr->u.sub.stride < 0){
@@ -476,10 +476,10 @@ static NhlErrorTypes MultiDVal_list_md_WriteSection
                                         sel_ptr->u.sub.stride = 1;
                                 }
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.start 
 					- sel_ptr->u.sub.finish))
-					/(float)fabs(((float)sel_ptr->u.sub.stride))) + 1;
+					/(double)fabs(((double)sel_ptr->u.sub.stride))) + 1;
 
 /*
 * Need to be able to determine which type of comparision < or > is needed to
@@ -502,10 +502,10 @@ static NhlErrorTypes MultiDVal_list_md_WriteSection
                                         sel_ptr->u.sub.stride = 1;
                                 }
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.finish 
 					- sel_ptr->u.sub.start))
-					/((float)sel_ptr->u.sub.stride)) + 1;
+					/((double)sel_ptr->u.sub.stride)) + 1;
 
 				if(sel_ptr->u.sub.stride < 0){
                                         compare_sel[i] = -1;
@@ -777,10 +777,10 @@ static NhlErrorTypes MultiDVal_list_s_WriteSection
 
                                 }
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.start 
 					- sel_ptr->u.sub.finish))
-					/(float)fabs(((float)sel_ptr->u.sub.stride))) + 1;
+					/(double)fabs(((double)sel_ptr->u.sub.stride))) + 1;
 
 /*
 * Need to be able to determine which type of comparision < or > is needed to
@@ -804,10 +804,10 @@ static NhlErrorTypes MultiDVal_list_s_WriteSection
 
                                 }
 
-				n_elem = (int)(((float)
+				n_elem = (int)(((double)
 					(sel_ptr->u.sub.finish 
 					- sel_ptr->u.sub.start))
-					/((float)sel_ptr->u.sub.stride)) + 1;
+					/((double)sel_ptr->u.sub.stride)) + 1;
 
 				if(sel_ptr->u.sub.stride < 0){
                                         compare_sel[i] = -1;
@@ -1147,10 +1147,10 @@ NclSelectionRecord *from_selection;
 
                                 }
 
-				n_elem_target = (int)(((float)
+				n_elem_target = (int)(((double)
 					(to_sel_ptr->u.sub.start 
 					- to_sel_ptr->u.sub.finish))
-					/(float)fabs(((float)to_sel_ptr->u.sub.stride))) + 1;
+					/(double)fabs(((double)to_sel_ptr->u.sub.stride))) + 1;
 
 /*
 * Need to be able to determine which type of comparision < or > is needed to
@@ -1173,10 +1173,10 @@ NclSelectionRecord *from_selection;
                                         to_sel_ptr->u.sub.stride = 1;
                                 }
 
-				n_elem_target = (int)(((float)
+				n_elem_target = (int)(((double)
 					(to_sel_ptr->u.sub.finish 
 					- to_sel_ptr->u.sub.start))
-					/((float)to_sel_ptr->u.sub.stride)) + 1;
+					/((double)to_sel_ptr->u.sub.stride)) + 1;
 
 				if(to_sel_ptr->u.sub.stride < 0){
                                         to_compare_sel[i] = -1;
@@ -1260,10 +1260,10 @@ NclSelectionRecord *from_selection;
 
                                 }
 
-				n_elem_value = (int)(((float)
+				n_elem_value = (int)(((double)
 					(from_sel_ptr->u.sub.start 
 					- from_sel_ptr->u.sub.finish))
-					/(float)fabs(((float)from_sel_ptr->u.sub.stride))) + 1;
+					/(double)fabs(((double)from_sel_ptr->u.sub.stride))) + 1;
 
 /*
 * Need from be able from determine which type of comparision < or > is needed from
@@ -1287,10 +1287,10 @@ NclSelectionRecord *from_selection;
 					return(NhlFATAL);
 				}
 
-				n_elem_value = (int)(((float)
+				n_elem_value = (int)(((double)
 					(from_sel_ptr->u.sub.finish 
 					- from_sel_ptr->u.sub.start))
-					/((float)from_sel_ptr->u.sub.stride)) + 1;
+					/((double)from_sel_ptr->u.sub.stride)) + 1;
 				from_compare_sel[i] = -2;
 			}
 			if((from_sel_ptr->u.sub.start > value_md->multidval.dim_sizes[from_sel_ptr->dim_num] - 1)||(from_sel_ptr->u.sub.start < 0)) {
